+++ core/cmd/app.go
@@ -256,6 +256,21 @@ func NewApp(client *Client) *cli.App {
 									Name:  "evmChainID",
 									Usage: "Chain ID for the key. If left blank, default chain will be used.",
 								},
+								cli.Uint64Flag{
+									Name:  "maxGasPriceGWei",
+									Usage: "Optional maximum gas price (GWei) for the creating key.",
+								},
+							},
+						},
+						{
+							Name:   "update",
+							Usage:  "Update the existing key's parameters",
+							Action: client.UpdateETHKey,
+							Flags: []cli.Flag{
+								cli.Uint64Flag{
+									Name:  "maxGasPriceGWei",
+									Usage: "Maximum gas price (GWei) for the specified key.",
+								},
 							},
 						},
 						{
@@ -382,6 +397,32 @@ func NewApp(client *Client) *cli.App {
 							Name:   "list",
 							Usage:  format(`List available CSA keys`),
 							Action: client.ListCSAKeys,
+						},
+						{
+							Name:  "import",
+							Usage: format(`Imports a CSA key from a JSON file.`),
+							Flags: []cli.Flag{
+								cli.StringFlag{
+									Name:  "oldpassword, p",
+									Usage: "`FILE` containing the password used to encrypt the key in the JSON file",
+								},
+							},
+							Action: client.ImportCSAKey,
+						},
+						{
+							Name:  "export",
+							Usage: format(`Exports an existing CSA key by its ID.`),
+							Flags: []cli.Flag{
+								cli.StringFlag{
+									Name:  "newpassword, p",
+									Usage: "`FILE` containing the password to encrypt the key (required)",
+								},
+								cli.StringFlag{
+									Name:  "output, o",
+									Usage: "`FILE` where the JSON file will be saved (required)",
+								},
+							},
+							Action: client.ExportCSAKey,
 						},
 					},
 				},