+++ core/cmd/eth_keys_commands_test.go
@@ -29,26 +29,28 @@ func TestEthKeysPresenter_RenderTable(t *testing.T) {
 	t.Parallel()
 
 	var (
-		address     = "0x5431F5F973781809D18643b87B44921b11355d81"
-		ethBalance  = assets.NewEth(1)
-		linkBalance = assets.NewLinkFromJuels(2)
-		isFunding   = true
-		createdAt   = time.Now()
-		updatedAt   = time.Now().Add(time.Second)
-		bundleID    = cltest.DefaultOCRKeyBundleID
-		buffer      = bytes.NewBufferString("")
-		r           = cmd.RendererTable{Writer: buffer}
+		address        = "0x5431F5F973781809D18643b87B44921b11355d81"
+		ethBalance     = assets.NewEth(1)
+		linkBalance    = assets.NewLinkFromJuels(2)
+		isFunding      = true
+		createdAt      = time.Now()
+		updatedAt      = time.Now().Add(time.Second)
+		maxGasPriceWei = utils.NewBigI(12345)
+		bundleID       = cltest.DefaultOCRKeyBundleID
+		buffer         = bytes.NewBufferString("")
+		r              = cmd.RendererTable{Writer: buffer}
 	)
 
 	p := cmd.EthKeyPresenter{
 		ETHKeyResource: presenters.ETHKeyResource{
-			JAID:        presenters.NewJAID(bundleID),
-			Address:     address,
-			EthBalance:  ethBalance,
-			LinkBalance: linkBalance,
-			IsFunding:   isFunding,
-			CreatedAt:   createdAt,
-			UpdatedAt:   updatedAt,
+			JAID:           presenters.NewJAID(bundleID),
+			Address:        address,
+			EthBalance:     ethBalance,
+			LinkBalance:    linkBalance,
+			IsFunding:      isFunding,
+			CreatedAt:      createdAt,
+			UpdatedAt:      updatedAt,
+			MaxGasPriceWei: *maxGasPriceWei,
 		},
 	}
 
@@ -62,6 +64,7 @@ func TestEthKeysPresenter_RenderTable(t *testing.T) {
 	assert.Contains(t, output, strconv.FormatBool(isFunding))
 	assert.Contains(t, output, createdAt.String())
 	assert.Contains(t, output, updatedAt.String())
+	assert.Contains(t, output, maxGasPriceWei.String())
 
 	// Render many resources
 	buffer.Reset()
@@ -75,6 +78,7 @@ func TestEthKeysPresenter_RenderTable(t *testing.T) {
 	assert.Contains(t, output, strconv.FormatBool(isFunding))
 	assert.Contains(t, output, createdAt.String())
 	assert.Contains(t, output, updatedAt.String())
+	assert.Contains(t, output, maxGasPriceWei.String())
 }
 
 func TestClient_ListETHKeys(t *testing.T) {
@@ -141,7 +145,9 @@ func TestClient_CreateETHKey(t *testing.T) {
 
 	set = flag.NewFlagSet("test", 0)
 	set.String("evmChainID", "", "")
+	set.Uint64("maxGasPriceGWei", 0, "")
 	c = cli.NewContext(nil, set, nil)
+	set.Set("maxGasPriceGWei", "12345")
 	set.Parse([]string{"-evmChainID", id.String()})
 	assert.NoError(t, client.CreateETHKey(c))
 
@@ -154,6 +160,44 @@ func TestClient_CreateETHKey(t *testing.T) {
 	// https://app.shortcut.com/chainlinklabs/story/17044/chainset-should-update-chains-when-nodes-are-changed
 	// states, err := app.KeyStore.Eth().GetStatesForChain(id)
 	// require.Len(t, states, 1)
+}
+
+func TestClient_UpdateETHKey(t *testing.T) {
+	t.Parallel()
+
+	ethClient, assertMocksCalled := newEthMock(t)
+	defer assertMocksCalled()
+	ethClient.On("BalanceAt", mock.Anything, mock.Anything, mock.Anything).Return(big.NewInt(42), nil)
+	ethClient.On("GetLINKBalance", mock.Anything, mock.Anything).Return(assets.NewLinkFromJuels(42), nil)
+	app := startNewApplication(t,
+		withKey(),
+		withMocks(ethClient),
+		withConfigSet(func(c *configtest.TestGeneralConfig) {
+			c.Overrides.EVMDisabled = null.BoolFrom(false)
+			c.Overrides.GlobalEvmNonceAutoSync = null.BoolFrom(false)
+			c.Overrides.GlobalBalanceMonitorEnabled = null.BoolFrom(false)
+		}),
+	)
+	ethKeyStore := app.GetKeyStore().Eth()
+	client, _ := app.NewClientAndRenderer()
+
+	// Create the key
+	key, err := ethKeyStore.Create(&cltest.FixtureChainID)
+	require.NoError(t, err)
+
+	// Update the key
+	set := flag.NewFlagSet("test", 0)
+	set.Uint64("maxGasPriceGWei", 0, "")
+	set.Set("maxGasPriceGWei", "12345")
+	set.Parse([]string{key.Address.Hex()})
+	c := cli.NewContext(nil, set, nil)
+	require.NoError(t, client.UpdateETHKey(c))
+
+	// Checking updated config
+	chain, err := app.ChainSet.Get(&cltest.FixtureChainID)
+	require.NoError(t, err)
+	price := chain.Config().KeySpecificMaxGasPriceWei(key.Address.Address())
+	require.Equal(t, assets.GWei(12345), price)
 }
 
 func TestClient_DeleteETHKey(t *testing.T) {