+++ core/web/config_controller_test.go
@@ -1,13 +1,13 @@
 package web_test
 
 import (
+	"github.com/smartcontractkit/chainlink/core/config"
 	"math/big"
 	"net/http"
 	"testing"
 	"time"
 
 	"github.com/smartcontractkit/chainlink/core/services/eth"
-	"github.com/smartcontractkit/chainlink/core/store/presenters"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/smartcontractkit/chainlink/core/internal/cltest"
@@ -25,7 +25,7 @@ func TestConfigController_Show(t *testing.T) {
 	defer cleanup()
 	cltest.AssertServerResponse(t, resp, http.StatusOK)
 
-	cp := presenters.ConfigPrinter{}
+	cp := config.ConfigPrinter{}
 	require.NoError(t, cltest.ParseJSONAPIResponse(t, resp, &cp))
 
 	assert.Contains(t, cp.RootDir, "/tmp/chainlink_test/")